Added the PDF version, and updated the index.html.
I have added the pdf version of the DressCode_CV_B421016 and updated the index.html file to include the link to the pdf file.

Please note that ppt was already submitted, and merged to the main branch. I wasn't aware about making changes to the index file, so I am doing it now.